删除和修改的功能和后台交互可以正常使用，具体还需要修改

+++ src/main/java/com/yan/ssm/service/UseService.java
@@ -20,12 +20,12 @@ public interface UseService {
 
     int updateSelective(User user);
 
-    int deleteById(Long id);
+    int deleteById(int id);
 
     List<User> selectByParams (Map<String,Object>params);
 
 
-    User selectByPrimaryKey(Long id);
+    User selectByPrimaryKey(int id);
 
     User selectByNick(String username) ;
 }
